+++ tests/test_fuzzy.lua
@@ -195,37 +195,27 @@ end
 
 T['process_lsp_items()'] = new_set()
 
-local new_item = function(newText, insertText, label)
-  return { textEdit = { newText = newText }, insertText = insertText, label = label }
+local new_item = function(filterText, label, newText, insertText)
+  return { filterText = filterText, label = label, textEdit = { newText = newText }, insertText = insertText }
 end
 
 local process_lsp_items = function(...) return child.lua_get('MiniFuzzy.process_lsp_items(...)', { ... }) end
 
 T['process_lsp_items()']['works'] = function()
-  local items
-
-  items = { new_item('___a', nil, nil), new_item('__a', nil, nil), new_item('_a', nil, nil) }
-  eq(process_lsp_items(items, 'a'), { items[3], items[2], items[1] })
-
-  items = { new_item(nil, '___a', nil), new_item(nil, '__a', nil), new_item(nil, '_a', nil) }
-  eq(process_lsp_items(items, 'a'), { items[3], items[2], items[1] })
-
-  items = { new_item(nil, nil, '___a'), new_item(nil, nil, '__a'), new_item(nil, nil, '_a') }
-  eq(process_lsp_items(items, 'a'), { items[3], items[2], items[1] })
+  local validate = function(items) eq(process_lsp_items(items, 'a'), { items[3], items[2], items[1] }) end
+  validate({ new_item('___a'), new_item('__a'), new_item('_a') })
+  validate({ new_item(nil, '___a'), new_item(nil, '__a'), new_item(nil, '_a') })
 end
 
 T['process_lsp_items()']['correctly extracts candidate from fields'] = function()
-  local items
+  -- filterText > label; no other fields should matter
+  local validate = function(items) eq(process_lsp_items(items, 'a'), { items[2], items[1] }) end
 
-  -- textEdit.newText > insertText > label
-  items = { new_item('__a', '_a', nil), new_item('_a', '__a', nil) }
-  eq(process_lsp_items(items, 'a'), { items[2], items[1] })
-
-  items = { new_item('__a', nil, '_a'), new_item('_a', nil, '__a') }
-  eq(process_lsp_items(items, 'a'), { items[2], items[1] })
-
-  items = { new_item(nil, '__a', '_a'), new_item(nil, '_a', '__a') }
-  eq(process_lsp_items(items, 'a'), { items[2], items[1] })
+  validate({ new_item('__a', '_a'), new_item('_a', '__a') })
+  validate({ new_item('__a', nil, '_a'), new_item('_a', nil, '__a') })
+  validate({ new_item(nil, '__a', '_a'), new_item(nil, '_a', '__a') })
+  validate({ new_item('__a', nil, nil, '_a'), new_item('_a', nil, nil, '__a') })
+  validate({ new_item(nil, '__a', nil, '_a'), new_item(nil, '_a', nil, '__a') })
 end
 
 T['get_telescope_sorter()'] = new_set()